Add functionallity for dynamic pdf files
The message "This document enabled extended features in Adobe Reader. The document has been changed since it was created and use of extended features is no longer available. Please contact the author for the original version of this document." can not be avoided, if the read only option is used on a XFA pdf file.